Seperated the components and hooks into different directories for best practices in file management
