+++ src/components/MoviesList.js
@@ -1,0 +1,10 @@
+import Movie from "./Movie";
+export default function MoviesList({ movies, onSelectMovie }) {
+  return (
+    <ul className="list list-movies">
+      {movies?.map((movie) => (
+        <Movie movie={movie} key={movie.imdbID} onSelectMovie={onSelectMovie} />
+      ))}
+    </ul>
+  );
+}
