+++ src/components/Results.js
@@ -1,0 +1,7 @@
+export default function Results({ movies }) {
+  return (
+    <p className="num-results">
+      Found <strong>{movies.length}</strong> results
+    </p>
+  );
+}
