+++ src/components/WatchedMoviesList.js
@@ -1,0 +1,15 @@
+import WatchedMovie from "./WatchedMovie";
+
+export default function WatchedMoviesList({ watched, onDeleteWatched }) {
+  return (
+    <ul className="list">
+      {watched.map((movie) => (
+        <WatchedMovie
+          movie={movie}
+          key={movie.imdbID}
+          onDeleteWatched={onDeleteWatched}
+        />
+      ))}
+    </ul>
+  );
+}
